scheduler top refactoring
