+++ kyo-scheduler/jvm/src/main/scala/kyo/scheduler/top/Printer.scala
@@ -1,103 +1,9 @@
-package kyo.scheduler.util
+package kyo.scheduler.top
 
-import java.lang.management.ManagementFactory
-import javax.management.Attribute
-import javax.management.MBeanServer
-import javax.management.ObjectName
-import javax.management.StandardMBean
-import javax.management.remote.JMXConnectorFactory
-import javax.management.remote.JMXServiceURL
-import kyo.scheduler.InternalTimer
-import kyo.scheduler.Scheduler
-import kyo.scheduler.WorkerStatus
-import scala.annotation.nowarn
-import scala.concurrent.duration.*
-import scala.io.StdIn
+object Printer {
 
-class Top(
-    status: () => Scheduler.Status,
-    enableTopJMX: Boolean,
-    enableTopConsoleMs: Int,
-    timer: InternalTimer
-) extends TopMBean {
+    def apply(status: Status) = {
 
-    private val mBeanServer: MBeanServer = ManagementFactory.getPlatformMBeanServer
-    private val objectName               = new ObjectName("kyo.scheduler:type=Top")
-
-    private var lastConsoleStatus: Scheduler.Status = null
-
-    if (enableTopConsoleMs > 0)
-        timer.schedule(enableTopConsoleMs.millis) {
-            val currentStatus = status()
-            if (lastConsoleStatus ne null) {
-                println(Top.print(currentStatus - lastConsoleStatus))
-            }
-            lastConsoleStatus = currentStatus
-        }
-
-    if (enableTopJMX) {
-        close()
-        mBeanServer.registerMBean(new StandardMBean(this, classOf[TopMBean]), objectName)
-    }
-
-    def getStatus() = status()
-
-    def close(): Unit =
-        if (mBeanServer.isRegistered(objectName))
-            mBeanServer.unregisterMBean(objectName)
-}
-
-trait TopMBean {
-    def getStatus(): Scheduler.Status
-}
-
-@nowarn
-object Top extends App {
-
-    lazy val (host, port) =
-        args.toList match {
-            case Nil =>
-                ("localhost", 1099)
-            case (host: String) :: Nil =>
-                (host, 1099)
-            case (host: String) :: (port: String) :: Nil =>
-                (host, port.toInt)
-            case args =>
-                throw new IllegalArgumentException("Expected host and port but got: " + args)
-        }
-
-    private lazy val jmxServiceURL         = new JMXServiceURL(s"service:jmx:rmi:///jndi/rmi://$host:$port/jmxrmi")
-    private lazy val jmxConnector          = JMXConnectorFactory.connect(jmxServiceURL)
-    private lazy val mBeanServerConnection = jmxConnector.getMBeanServerConnection
-
-    private lazy val objectName = new ObjectName("kyo.scheduler:type=Top")
-
-    private var lastStatus: Scheduler.Status = null
-
-    delayedInit {
-        try {
-            while (true) {
-
-                val currentStatus = mBeanServerConnection.getAttribute(objectName, "Status").asInstanceOf[Scheduler.Status]
-
-                if (lastStatus != null) {
-                    val diffStatus = currentStatus - lastStatus
-                    val clear      = "\u001b[2J\u001b[H"
-                    println(clear + "\n" + print(diffStatus))
-                }
-
-                lastStatus = currentStatus
-                Thread.sleep(1000) // Refresh every 1 second
-            }
-        } catch {
-            case e: Exception =>
-                e.printStackTrace()
-        } finally {
-            jmxConnector.close()
-        }
-    }
-
-    private[util] def print(status: Scheduler.Status): String = {
         val sb = new StringBuilder()
 
         sb.append(f"""
