+++ kyo-scheduler/shared/src/main/scala/kyo/scheduler/Task.scala
@@ -27,20 +27,9 @@ trait Task {
             if (state < 0) -state + v
             else state + v
     }
-
-    def status(): Task.Status =
-        Task.Status(
-            shouldPreempt(),
-            runtime()
-        )
 }
 
 object Task {
-
-    case class Status(
-        preempting: Boolean,
-        runtime: Int
-    )
 
     implicit val taskOrdering: Ordering[Task] =
         new Ordering[Task] {
